feat: widen exact linear rows and the LP relaxation to Long coefficients

+++ klause/src/commonMain/kotlin/com/eignex/klause/lp/RelaxationBuilder.kt
@@ -18,7 +18,7 @@ interface RelaxationBuilder {
     fun linearRow(
         op: LinearOp,
         intVars: IntArray,
-        coeffs: IntArray,
+        coeffs: LongArray,
         bound: Long,
         contribution: Contribution = Contribution.CORE,
     )
